// apps/www/src/interface.ts
export interface ITeam {
  gender: string
  slug: string
  id: string
  team_id: string
  name: string
  short_name: string
  abbreviation: string
  icon: string
  qualifier: string
  statistics: ITeamStatistics
  players: IPlayer[]
  form: string[]
}
// ...
export interface ISportEventContext {
  sport: ISport
  season: ISeason
  competition: ILeague
  category: ICategory
}

export interface ISportEvent {
  id: string
  sport_event_id?: string
  start_time: string
  start_time_formatted: string
  start_date_formatted: string
  competitors: ITeam[]
  sport_event_context: ISportEventContext
}

export type ISportEventStatusType = 'not_started' | 'closed' | 'live'

export type ISportEventMatchStatusType =
  | 'not_started'
  | 'ended'
  | '1st_half'
  | '2nd_half'
  | 'overtime'

export interface ISportEventStatus {
  clock: {
    played: string
  }
  start_time_formatted: string
  home_score: number
  away_score: number
  status: ISportEventStatusType
  match_status: ISportEventMatchStatusType
  winner_id?: string
  match_tie?: boolean
}

export interface ISportEventStatistics {
  totals: {
    competitors: ITeam[]
  }
}

export interface ISportEventSummary {
  id: string
  sport_event_status: ISportEventStatus
  sport_event: ISportEvent
  statistics: ISportEventStatistics
}